% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\docType{data}
\name{diamonds_fake}
\alias{diamonds_fake}
\title{3 variables going mad}
\format{
A data frame with 100 rows and 3 variables:

\describe{
\item{x1}{uniform distributed}
\item{x2}{normal distributed}
\item{y}{binomial distributed}
}
}
\source{
simulation and motivated by \url{http://www.diamondse.info/}
}
\usage{
data(diamonds_fake)
}
\description{
A dataset containing blub
}
\keyword{datasets}
