% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func01.R
\name{fnc}
\alias{fnc}
\title{Elementwise Square}
\usage{
fnc(x, y)
}
\arguments{
\item{x}{A numeric vector (or scalar).}

\item{y}{Adds a constant to the squares of the numeric vector}
}
\value{
The output is a vector with the same length as code{x}.
}
\description{
code{ fnc } returns square of each element of a numeric vector
}
\examples{
fnc (c(1:5), 3)
\dontrun{
 fnc(1:10000, -2)}

}
\references{
Klauer , K. C. (2010). Hierarchical multinomial processing tree models: A latent trait approach.
\emph{Psychometrika , 75(1)}, 70 98.
doi: \doi{10.1007/s11336-009-9141-0}

Klauer , K. C. (2010). Hierarchical multinomial processing tree models: A latent trait approach.
\emph{Psychometrika , 75(1)}, 70 98.
doi: \doi{10.1007/s11336-009-9141-0}
}
\author{
Julius Fenn

Katja Pollak
}
