% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func02.R
\name{fnc2}
\alias{fnc2}
\title{Elementwise divide by X}
\usage{
fnc2(x, y)
}
\arguments{
\item{x}{A numeric vector (or scalar).}

\item{y}{division}
}
\value{
The output is a vector with the same length as code{x}.
}
\description{
code{ fnc2 } returns division  of each element of a numeric vector
}
\examples{
fnc2(c(1:5), 3)
\dontrun{
 fnc2(1:10000, 5)}

}
\references{
Klauer , K. C. (2010). Hierarchical multinomial processing tree models: A latent trait approach.
\emph{Psychometrika , 75(1)}, 70 98.
doi: \doi{10.1007/s11336-009-9141-0}
}
\seealso{
\code{\link{fnc}}
}
\author{
Julius Fenn
}
