% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pck01.R
\docType{package}
\name{pck01}
\alias{pck01}
\title{pkgname: My frist package}
\description{
The package pkgname provides two functions and a dataset:

\code{\link{fnc}} a function to calculate the square

fnc2 a function to calculate a division

dat a data frame with 3 random variables
}
\section{Functions}{

The functions ...
}

\section{Datasets}{

The datasets ...
}

